Restructured Link component structure
